Fix conflicts with older WP tests
Recently the tests have changed the way that the factory is created,
and it is offered as a magic property. Because of this, it was causing
a conflict with the newer tests when we defined that property. The
older version of the test suite conflicted with our fix however, so the
solution is just not to define this property so that the magic can work.