Split registry getters
This makes it easier for static analysis tools like scrutenizer to
determine when only an array will be returned, and when it is possible
that false will be returned. It also makes the methods simpler and
decreases the likelihood of unexpected bugs by inadvertently passing
null for one of the values and getting an array instead of a single
object.