Tests for the hook args class
I’ve added a `_doing_it_wrong()` notice to the `ascend()` method when
already at the top of the hierarchy. That would generally only happen
when some logic is messed up somewhere. So I figured that this would be
a helpful notice. I also added a return value to the method, mostly for
the benefit of child methods.