Tests for the main hook event firer
I’ve removed the `$slug` because it is currently unused, and appears to
be unneeded. Also, to be of any use, we’d need to add a `get_slug()`
method on the interface, and possibly extend the spam and reverse
firers  from the main firer (they currently implement the interface
directly).

I’ve simultaneously added a `get_slug()` method to the main extension
class, since that very well could prove useful.