Don't push fields onto validator when getting args from hierarchy
This was the incorrect behavior.

The field on the validator corresponds to a field in the UI. In the
case where we are getting an arg from a hierarchy, that hierarchy is
supplied by a single field in the UI. So we shouldn’t be treating it as
if it is a sub-field.

A result of the incorrect behavior would be that the fields would not
only be pushed on when they shouldn’t be, but they would never be
popped off. This would lead to the field stack becoming bloated and
desynchronized.