Tests for the hook reaction validator
I’ve removed the check for whether the hook args are empty, because now
that `get_all()` always returns an array it wasn’t working anyway
(since it was strictly checking for boolean `false`). I don’t think
there is any application for arg-less events, however, the fact that
the expected args aren’t registered will be made apparent soon enough.

I’ve also added a `get_settings()` method, even though I guess it is
essentially a duplicate of `get_all_meta()`. It is a more obvious name
when you are thinking in terms of “settings”.