Pass the storage object to reaction objects
We were actually already doing this, but it wasn’t documented in the
reaction interface. Now I’ve removed the $reactor_slug and
$network_wide args in favor of just deriving them from the storage
object. (And I’ve introduced the methods needed to access them.)

At the same time I’ve updated things so that the reaction classes are
only constructed in a single place. I’ve also made the get_reaction()
method check that the reaction exists, now that the reaction object
will no longer have an exists() method.

See #12