Don't assume event slug will be stored as reaction meta
In addition, we stop storing the event slug as metadata for the option
storage method, and use the value in the index instead. This avoids
having to synchronize the value in both places.

I’ve also changed the reaction index used by the option storage method
to be indexed by the reaction ID. This greatly simplifies retrieving
the data for a specific reactor from the index.

Fixes #37