Tests for the main reaction object
I decided to remove the option to pass a reaction object as the `$id`
parameter of the constructor. This was a vestige carried over from the
`WP_Post` class, but it doesn’t really provide any benefit due to the
way that the reactions are constructed. The storage class basically
acts as a factory for them, so they are never constructed outside of
it. And more special handling would be required within each of the
storage class methods to actually make this useful—and even then it
wouldn’t be very useful at that.