Tests for the options reaction object class
I’ve changed the `get_meta()` method to return `false` instead of
`null` when the metadata isn’t found, because it is on par with
WordPress’s meta API. The benefit of `null` is that it may be easier to
differentiate from values that are simply stored a boolean in the
database. However, it would likely be difficult to actually return
`null` instead of `false` at the correct times if the WordPress
metadata API was used in an implementation. Also, the lack of parity
means that it would be easy for folks to run into bugs because they
just assumed that `false` would be returned.