Tests for the conditions extension class
In the process I’ve made it so that we no longer discard invalid
conditions during validation. I don’t remember the exact rationale
behind that, but I think it was sort of with the intention to return a
valid result no matter what is passed in. That ideal doesn’t really
carry over into cases when we have required values and no defaults, so
I’m not going to pursue it. There is no reason that we should mangle
the settings array just because some parts are invalid.

Also in the process a bug was fixed that would cause the args hierarchy
not to be restored to the default state when an unmet condition was
encountered. I’m not sure that this is really a big deal anyway,
because in that case we would be done checking whether the reactor
should fire. But I suppose if we introduce an after miss method at some
point, it will be good to know that the args hierarchy will not be left
in an odd state.