Move UI script data setup to its own function
Instead of it being a method on the points type screen class.

I’ve toyed with the idea of adding filters that would let us only
enqueue the scripts and data for certain extensions and rectors,
instead of all of them. However, if we don’t enqueue everything but
later decide that we need to (i.e., in the footer), we’ll run into
trouble. So for now we’ll just enqueue everything.