Standardize validate_settings() method return value
The method is intended to return the validated settings. If the
settings are invalid, any errors will be added to the validator, so we
don’t need to return `false` to indicate that the settings are invalid.
Instead we just return the settings anyway, even though they might not
be valid.