Tests for the entity array condition
In the process, I’ve modified the event args class to fail gracefully
when no validator is set. Although it is intended to be used with a
validator most of the time, when we are checking if the sub-conditions
are met (the conditions on the items in the entity array), they have
already been validated, so we don’t actually need a validator. We can’t
easily create one just for this either, because the validator now needs
to be constructed with a reactor object (which we don’t have available
to us in this case).

I’ve moved some condition-related data providers to the main hooks test
case so that they can be reused.