Register current site arg for all available hooks
This might make you think that we could just make this a special arg
that was always available for every even on multisite. However, some
events may be in context of the network as a whole instead of a
particular site. Because of that this arg needs to be added manually to
the events that can take place in the context of a particular site.