Remove .button-link from add/cancel condition buttons
The button link styling will be changing in WordPress 4.4, and it no
longer fits these buttons with text.

We retain the styles for the icon-only buttons (delete condition and
add condition), and they continue to look fine thanks to addition
padding we had for them.

I’m not sure why the horizontal padding was 0 for these buttons, and I
guess we could have just changed that styling for the other two buttons
instead of going back to non-link buttons. But I think this better
matches the rest of the styles, both in WordPoints and in WordPress
core.