Use \ instead of - to delimit dynamic slugs
Because we need to register entities (and therefore hook events and
args, etc.) dynamically for each post type, we have to give them
dynamic slugs. To avoid naming conflicts, I was prefacing the slugs
with a generic string (like `post`) and then adding a hyphen (`-`) and
then the dynamic portion of the string. However, I think that the
backslash (`\`) provides a better delimiter, because it makes it more
obvious that we have a generic prefix or “namespace” and then a dynamic
portion.

I’ve also updated the tests for the dynamic hook args and the dynamic
entity relationships so that they test for the case when multiple
delimiters are used in the string. (In that case, the string should be
broken only at the first delimiter.)

See #42