grunt: stop hacking the watch watch task for browserify
I think that actually this didn’t make much difference in how long the
task took to complete. The reason for removing it was that for some
reason I haven’t been able to get browserify to turn off the
`full-paths` flag, which was causing the full file paths to be included
in the built JS files. Removing this hack fixes that issue, and who
knows what else.