Pass the reaction around in the period extension
We were using the extension slug instead of the event slug in two
places. We couldn’t get the event slug because we only had the reaction
ID, instead of the reaction object. We could have just passed the event
slug too, but we might need other data from the reaction in the future.

See #10