Completed basic periods extension
Prior to now we had some parts of a basic periods extension, but it
wasn’t complete. We’ve now added an event hit listener to actually
record a period in the database when the target is hit.

The hit listener actually didn’t exist before, so that’s a new
addition. Along the way we’ve also made some other changes, such as:

- Storing the reaction ID in the database instead of a “hook ID”
composed of the reaction ID and event slug. We don’t need to take the
event into account since reactions don’t usually change from one event
to the next. Behavior for when that happens is currently undefined.
- Store the period signature (composed of the hash of a JSON-encoded
single-dimensional array of the values of the event args the period is
related to, keyed by the imploded arg “paths”) in a fixed-length field,
since a SHA256 hash will always be 64 characters long. This may be
beneficial when performing queries.
- Store the hit time instead of an expiration date, as this allows us
to move more easily between different period settings without strange
things happening.
- The code to calculate whether the period was over if one was found in
the database was fixed so that it now works correctly for both relative
and absolute periods. We default to absolute periods.

Includes a whole passel of unit tests, which cover behavior for most
edge cases.

Fixes #10