Change to a date time field for period hit time
We use a date time field for the points logs table, so I’ll use it here
too to be consistent. Also, we were using a `BIGINT` field to store the
timestamp anyway, which isn’t correct.

Using a date time field means that we avoid issues with the Y2038 bug,
and don’t have to worry about conversion between timezones that MySQL
does to the timestamp field. The datestamp field is timezoneless, so we
don’t have to worry about the original timezone, it will always be
interpreted relative to the current one.

See #10