Add period signature key to the periods table
The key is on the `reaction_id` and `signature` fields, because we
query on then when searching for the last hit for a particular period.

I’ve made these not null because it simplifies things for MySQL when it
is deciding how to use the indexes, and we never needed these to be
null since they are vital pieces of data.

We only use the first 8 characters of the reaction signature in the
index to save disc space. There isn’t really a need to store the entire
signature, since it is basically a “random” string and is likely to be
different in the first few characters if it is different at all.

`EXPLAIN` tells me that this allows our query to be classified as a
`ref` join type. See
https://dev.mysql.com/doc/refman/5.5/en/explain-output.html#jointype_ref

See #10