Unit tests: convert database errors into errors
`$wpdb` contains a `print_error()` method that it calls whenever it
wants to tell the world about an error it encountered. Unfortunately,
it doesn’t actually call `trigger_error()` to trigger a real error.
Instead it prints the error to the screen and/or logs it with
`error_log()`. So when a database error is encountered during a test,
the error is logged and printed to the screen, but unless there are
other errors, the test will not fail. We fix that by hacking the error
array with an `ArrayAccess` object that triggers the error whenever a
new error is added to the array. Since new errors are added to the
array even when error suppression is on, this means that we can disable
`$wpdb`’s regular error handling.

Now when a database error occurs during a test, the test will be marked
as “errored”.

I happened to mock this together while toying with some things related
to #66.