Enable getting reaction GUIDs
A reaction’s GUID is determined based on the storage group to which it
belongs (standard vs network) and the ID of the exact context in which
it exists (the ID of the site and/or network), in addition to the
reactor slug and reaction ID.

This commit makes it possible to get the context ID and storage group
of a reaction from the reaction object.

This enables us to calculate the reaction GUID from the reaction
object. It is the first step toward possibly allowing custom reaction
types—reaction groups that correspond to other contexts.

See #57