Don't trigger intentionally suppressed DB errors
Sometimes code is intentionally suppressing database errors. When that
is the case, we don’t need to elevate them to PHP errors. But in all
other cases we do.

To accomplish that, we just stop setting `$wpdb->suppress_errors` to
`true`, leaving it as `false` unless it is changed intentionally by the
code being tested. Then in our error handler we check if suppression is
on, and don’t trigger a PHP error if it is. (We don’t have to turn off
`$wpdb->show_errors`, because the moment we trigger the error PHPUnit
will throw an exception ending execution along that path—meaning that
the error will not be logged or displayed like `$wpdb` would normally
do otherwise.)

See dba9e33