Make hooks network mode a generic mode
Previously, network mode was a boolean value, so that network mode was
either on or off. This changes it from being called network mode, to
just being called the current mode of the hooks API. The mode is now
identified by a string slug, which means it is possible for other modes
besides just `standard` and `network` to be added in the future.

This doesn’t change the behavior of how the default value for the mode
is determined based on the current context. What it does do is move the
code for checking whether we are in network context to a utility
function, that is now used by the entity context API as well as here in
the hooks API. This fixes #61 (see #63 also).

Within the reactor bootstrap, how the reaction storage objects are
accessed was unchanged, but the logic has been updated to be more
generic, allowing custom reaction storage objects to be added using new
properties.

One result of these changes is that it is no longer so dangerous to
change the current mode to `’standard’` when WordPoints has set it to
`network`. Before, this might cause reactions to be pulled from a
particular site on multisite when in network context. But now that
network context is no longer a deduction from the current mode, but is
determined entirely separately, we can set the current mode to
`standard` even in the network admin, and we’ll be unable to access the
standard reactions because the storage object will detect that it is
out of context. In other words, the current mode no longer dictates the
current context, and attempting to use a mode out of context will not
work.

See #57, #12