Introduce reaction group registry
Using this registry new reaction groups can be registered for a
reactor. The reaction groups should only be accessed through the
reactor itself, however, using one of two methods: 1) the current group
can be accessed through the `$reactions` property, 2) any group may be
accessed by using the `get_reaction_group()` method and passing in the
group slug. This replaces the need for the reaction groups to be
defined as properties of the reactor class.

One side-effect of this is that reaction storage objects are not longer
persistent from one access of the `$reactions` property to the next.
This required some changes in our tests, but otherwise shouldn’t be a
problem.

See #57