Specify contexts using single slugs not arrays
Because contexts are hierarchical, we were have been specifying
contexts using arrays in order from highest to lowest. However, all of
the contexts share a single namespace, so it can be misleading to treat
them as if they only identify a particular context when associated with
the parent slugs as well. We now specify the parent context in the
context class itself, so it is much DRYer just to define the parent
there, instead of all over the place like this.

In the process, I’ve entirely removed the mock network reaction storage
object, because it is unused.

This changes the order the get context ID utility function returns the
IDs in. Previously it returned them from highest to lowest, but now it
returns them from lowest to highest.

See #63