Introduce hook fire class
This allows us to encapsulate all of the data associated with a hook
fire. This will make it easier to add new data associated to a hook
fire in the future without having to constantly be increasing the
number of args that a bunch of different methods take.

This will pave the way for us to associate a hit ID and data from the
hit log database with each fire.

See #55, #68