Log hits in reverse firer
Also updates the firer so that it only fire when there is a prior hit
that could be reversed.

In the process, the firer and some `*reverse*()` methods had to be
refactored to accept a fire object, as has already been done with the
hit and fire methods.

To prevent hits from being reversed multiple times, the superseded
hit’s `superseded_by` field is updated in the database with the ID of
the reverse hit.

Fixes #68
See #55