Use hit logs table in periods extension
Instead of storing all of this information twice, we just join the
periods table on the hits table. This also updates the extension to
properly relate periods to the whole reaction GUID, instead of just the
ID. Since this relates the periods to the reaction context, it also
corrects the operation of the extension on multisite.

Fixes #84, #56
See #10