Disable network context when not on multisite
Because this function returns true during an Ajax request with a
referrer matching the network admin URL, when multisite was not enabled
it would return true for Ajax requests, since the `network_admin_url()`
function will just return the `admin_url()`.

This was causing a fatal error to occur when trying to save a reaction,
because the hooks API would be put into network mode, even though it
isn’t available when not on multisite. As a result attempting to call
the `$reactor->reactions->create_reaction()` method would be a call to
a method on a non-object.

Fixes #91