Centralize Ajax nonce creation for reactions
And use the entire GUID in the nonce action, or the entire context ID
for the `create` nonce. This may help make it possible to have
fine-grained control over who can modify what reactions. The nonce
handling shouldn’t be relied on to exclude users who shouldn’t have the
rights from modifying reactions, but this makes the handling more
robust when there are multiple different modes and contexts.

See #91