Don't double quote MySQL password
This was failing because the quotes were being added by
`escapeshellarg()`, so we didn’t need to add a second set.

Also, because the value was quoted, the empty value would actually be
equal to `”’’”`, and not `’’`.

Anyway, we leave out the password entirely for `mysqldump`, but leave
it in even if empty for `mysql`, so we can test to see if both ways
work. if they do, we’ll just pass the empty passwords to the commands,
since the logic is simpler.

See #92, WordPoints/hooks-api#46