Fix "this.model is undefined"
We were getting this error on the Points Types screen when each of the
reactors and extensions was being added via the `collection.add()`
method to its collection.

This was ultimately because the collections were being declared in the
wrong order.

Probably related to https://core.trac.wordpress.org/ticket/34350

Discovered while working on #103