Allow extensions to be firer-agnostic
In the process we had to update the conditions and periods extensions,
and they are now both firer-agnostic.

The PHPUnit tests also had to be updated, and in the process we
introduced a factory to create hook firers.

The JS needed to be updated as well, and this resulted in us having to
decide how firers should be handled in the UI. Previously the firer was
pretty much left out of the UI entirely. What we did was allow each
reactor to specify which firers it supported, but left the decision of
what firers actually were exposed in the UI up to the Reaction object.
Our core reaction object currently just displays settings for the
primary firer that the reactor lists.

The main remaining issue is that the Entity Array Contains condition
currently employs what is essentially a hack to make the settings
validation work. It continues to use the
`Conditions_Extension::validate_settings()` method, even though this
requires it to pass a firer-indexed array. The settings for the
sub-conditions don’t need to be firer-indexed though, and so currently
they aren’t stored that way. So what the code does is just grab a firer
slug from the firers registry and use that for the index. This means
that any errors will have incorrect field paths though. We’ll likely
revisit this.

Fixes #103