JS: Simplify condition hierarchy handling
Each condition has a two-part hierarchy. The first part is the
hierarchy of the set of condition groups as a whole, and the hierarchy
of each condition group within the set. Previously the information for
each of these was stored inconsistently, sometimes on the view,
sometimes on the model, and sometimes on both.

We now store this info only on the condition groups and condition group
models, not all over the place.

See #106