Improve entity array contains condition validation
Previously we were essentially using a hack to provide a firer slug and
use the `validate_settings()` method. This was not only messy, it also
had the effect of inserting the firer slug into the field stack, so
that the field would be wrong if there was an error. This is now fixed.

See #103