Expose entity attribute storage info
All of the attributes that we currently have are stored in an entity
field (or “attribute”). However, we are leaving the possibility open
here that attributes could be stored elsewhere in the future. Let’s
hope that we don’t live to regret that.

How the attribute storage data is exposed is currently dependent upon
the storage type of the attribute. The attribute must implement an
interface that exposes the information that is needed for that type of
storage.

See #124