Use a single method to get entity storage info
This will return an array that contains all of the storage information
for the entity. Much debate has raged about whether this approach
should be used or we should use interfaces that return objects. In the
end, this approach using a single method and returning an array has
prevailed. At least for now.

See #124