Pin the WP Browser version
The latest version is no longer working with the version of Codeception
that we are using (2.1.4). Probably we should update them both, but I
want to get this working again before I mess with that.

The error that currently occurs in Travis:

```
PHP Fatal error:  Class 'Codeception\Application' not found in
/home/travis/build/WordPoints/hooks-api/vendor/lucatume/wp-browser/wpcep
t on line 13

PHP Stack trace:

PHP   1. {main}()
/home/travis/build/WordPoints/hooks-api/vendor/lucatume/wp-browser/wpcep
t:0
```

See #92