Move event fire method from router to main app
This offers a more generic interface that is less strapped to the
WordPress action API.

In the process I also found that some to the router tests that had
before used a mock router were actually not really testing anything.
They were passing because they were asserting that 0 fires should have
occurred, but this wasn’t actually testing the code, since
`do_action()` was never called. This is now fixed.

Fixes #134