Always cache the number of hits
Otherwise we’ll run the query twice when there are no hits, once in
`should_hit()` and once in `after_miss()` (since we will ourselves
force a miss, so `after_miss()` will be called). This is because our
failed query checking code was too aggressive. We could have just
switched to  `is_array()` instead of `!`, but I decided to just
simplify things down to a single return point there.

See #99