Remove TODOs suggesting methods be removed
We’ll keep both of these methods around. Neither one is absolutely
necessary, but I see no reason to remove them at this point. It is true
that we don’t use `Entity::get_human_id()`, but that doesn’t mean that
we’ll never have a use for it.

See #136