Correct test for reaction ID incrementing
We copy/pasted this from the reaction store code, and forgot to change
`$this->slug` to the correct value. This silently succeeded locally,
but failed on older versions of WP in Travis. Another reason to hate
magic getters (which are part of the newer core test suite, and were
masking this).

See #136, #87