Fix admin apps
We were still using magic properties to try to access the admin apps,
but that isn’t supported anymore.

Includes unit tests for this method so that it would be more obvious to
us if we break it in the future. We have to include the WordPoints
admin bootstrap because otherwise `wordpoints_get_main_admin_menu()`
won’t be defined.

See #87