Only fire post type actions for correct post types
This splits the post- and comment-related actions per post type, and
ensures that the action for a particular post type only fires for posts
of that type.

We remove the old check for whether a post type is public within the
post publish/depublish actions, since these actions will now only be
firing for post types that they are specifically registered for. We
don’t currently want to register them for any non-public post types,
but this doesn’t need to be blocked by the action logic in case we want
to do that in the future.

Fixes #155