Introduce maybe-network option functions
These function allow you to return a regular or network option,
whichever one you need. The simplifies code that needs to toggle
between these two different types of options.

If the type of option is not specified, whether to return a network
option will be determined based on whether WordPoints is network active.

See WordPoints/wordpoints#305